Add new hook: before_logger
This can be used for example to filter before logging, before
the target system is actually modified, and so that the run
output only logs items that are actually processed.

If using before_commit as a filter stage, it doesn't indicate
the skip on the output in any way.

Better aproach could be to modify the filters to output a custom
a message whin filter hits, but this was way more straightforward.